Create README.md
# Game_of_life

The Game of Life is not your typical computer game. It is a 'cellular automaton', and was invented by Cambridge mathematician John Conway.

This code will generate a random matrix as the inital configuration of the world of the game. And I wrote two versions of the game with and without periodic boundary.

It's very interesting just to watch how those patterns evolve and think the philosophy behind the game.

Have fun!!!